+++ buchfink/datatypes.py
@@ -7,4 +7,4 @@ from rotkehlchen.assets.asset import Asset
 from rotkehlchen.chain.ethereum.trades import AMMTrade
 from rotkehlchen.exchanges.data_structures import Trade
 from rotkehlchen.fval import FVal
-from rotkehlchen.typing import TradeType
+from rotkehlchen.typing import TradeType, Timestamp